pending cards and clean up
