+++ packages/arb-token-bridge-ui/src/components/TransferPanel/WithdrawalCardConfirmed.tsx
@@ -9,6 +9,10 @@ import { Tooltip } from '../common/Tooltip'
 import { formatAmount } from '../../util/NumberUtils'
 import { useIsConnectedToArbitrum } from '../../hooks/useIsConnectedToArbitrum'
 import { sanitizeTokenSymbol } from '../../util/TokenUtils'
+import {
+  isCustomAddressTx,
+  CustomAddressTxExplorer
+} from '../TransactionHistory/TransactionsTable/TransactionsTable'
 
 export function WithdrawalCardConfirmed({ tx }: { tx: MergedTransaction }) {
   const { l2 } = useNetworksAndSigners()
@@ -50,6 +54,14 @@ export function WithdrawalCardConfirmed({ tx }: { tx: MergedTransaction }) {
             <span className="flex flex-nowrap gap-1 text-sm text-ocl-blue lg:text-base">
               L1 transaction: Will show after claiming
             </span>
+            {isCustomAddressTx(tx) && (
+              <span className="flex flex-nowrap gap-1 text-sm text-ocl-blue lg:text-base">
+                <CustomAddressTxExplorer
+                  tx={tx}
+                  explorerClassName="arb-hover text-blue-link"
+                />
+              </span>
+            )}
           </div>
         </div>
