+++ packages/arb-token-bridge-ui/src/components/TransferPanel/WithdrawalCardUnconfirmed.tsx
@@ -8,6 +8,10 @@ import { Tooltip } from '../common/Tooltip'
 import { formatAmount } from '../../util/NumberUtils'
 import { useMemo } from 'react'
 import { sanitizeTokenSymbol } from '../../util/TokenUtils'
+import {
+  isCustomAddressTx,
+  CustomAddressTxExplorer
+} from '../TransactionHistory/TransactionsTable/TransactionsTable'
 
 export function WithdrawalCardUnconfirmed({ tx }: { tx: MergedTransaction }) {
   const { l1, l2 } = useNetworksAndSigners()
@@ -47,6 +51,14 @@ export function WithdrawalCardUnconfirmed({ tx }: { tx: MergedTransaction }) {
             <span className="flex flex-nowrap gap-1 text-sm text-ocl-blue lg:text-base">
               L1 transaction: Will show after claiming
             </span>
+            {isCustomAddressTx(tx) && (
+              <span className="flex flex-nowrap gap-1 text-sm text-ocl-blue lg:text-base">
+                <CustomAddressTxExplorer
+                  tx={tx}
+                  explorerClassName="arb-hover text-blue-link"
+                />
+              </span>
+            )}
           </div>
         </div>
